+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/table/extractor/TablesExtractorTest.scala
@@ -1,28 +1,50 @@
 package pl.touk.nussknacker.engine.flink.table.extractor
 
 import org.apache.flink.table.api.DataTypes
-import org.apache.flink.types.Row
+import org.apache.flink.table.catalog.{Column, ResolvedSchema}
+import org.scalatest.LoneElement
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks
-import pl.touk.nussknacker.engine.api.typed.typing.Typed
+import pl.touk.nussknacker.engine.flink.table.utils.DataTypesConversions._
 import pl.touk.nussknacker.engine.flink.table.TableTestCases.SimpleTable
 import pl.touk.nussknacker.engine.flink.table._
-import pl.touk.nussknacker.engine.flink.table.extractor.TableExtractorTest.invalidSqlStatements
+import pl.touk.nussknacker.engine.flink.table.extractor.TablesExtractorTest.invalidSqlStatements
+import pl.touk.nussknacker.test.ValidatedValuesDetailedMessage
 
-class TableExtractorTest extends AnyFunSuite with Matchers with TableDrivenPropertyChecks {
+import scala.jdk.CollectionConverters._
+
+class TablesExtractorTest
+    extends AnyFunSuite
+    with Matchers
+    with LoneElement
+    with ValidatedValuesDetailedMessage
+    with TableDrivenPropertyChecks {
 
   test("extracts configuration from valid sql statement") {
     val statements        = SqlStatementReader.readSql(SimpleTable.sqlStatement)
-    val dataSourceConfigs = TableExtractor.extractTablesFromFlinkRuntime(statements)
-
-    val expectedResult = TableExtractorResult(
-      tableDefinitions = List(SimpleTable.tableDefinition),
-      sqlStatementExecutionErrors = List.empty
+    val tablesDefinitions = TablesExtractor.extractTablesFromFlinkRuntime(statements).validValue
+    val tableDefinition   = tablesDefinitions.loneElement
+    val sourceRowType     = tableDefinition.sourceRowDataType.toLogicalRowTypeUnsafe
+    sourceRowType.getFieldNames.asScala shouldBe List(
+      "someString",
+      "someVarChar",
+      "someInt",
+      "someIntComputed",
+      "file.name"
     )
+    sourceRowType.getTypeAt(0) shouldEqual DataTypes.STRING().getLogicalType
+    sourceRowType.getTypeAt(1) shouldEqual DataTypes.VARCHAR(150).getLogicalType
+    sourceRowType.getTypeAt(2) shouldEqual DataTypes.INT().getLogicalType
+    sourceRowType.getTypeAt(3) shouldEqual DataTypes.INT().getLogicalType
+    sourceRowType.getTypeAt(4) shouldEqual DataTypes.STRING().notNull().getLogicalType
 
-    dataSourceConfigs.sqlStatementExecutionErrors shouldBe empty
-    dataSourceConfigs shouldBe expectedResult
+    tableDefinition.sinkRowDataType.toLogicalRowTypeUnsafe.getFieldNames.asScala shouldBe List(
+      "someString",
+      "someVarChar",
+      "someInt",
+      "file.name"
+    )
   }
 
   test("extracts configuration from tables outside of builtin catalog and database") {
@@ -42,33 +64,26 @@ class TableExtractorTest extends AnyFunSuite with Matchers with TableDrivenPrope
        |);""".stripMargin
 
     val statements        = SqlStatementReader.readSql(statementsStr)
-    val extractionResults = TableExtractor.extractTablesFromFlinkRuntime(statements)
+    val tablesDefinitions = TablesExtractor.extractTablesFromFlinkRuntime(statements).validValue
 
-    extractionResults.sqlStatementExecutionErrors shouldBe empty
-    extractionResults.tableDefinitions shouldBe List(
-      TableDefinition(
-        tableName,
-        typingResult = Typed.record(Map("someString" -> Typed[String]), Typed.typedClass[Row]),
-        columns = List(
-          ColumnDefinition("someString", Typed[String], DataTypes.STRING())
-        )
-      )
+    tablesDefinitions.loneElement shouldBe TableDefinition(
+      tableName,
+      ResolvedSchema.of(Column.physical("someString", DataTypes.STRING()))
     )
   }
 
   test("returns errors for statements that cannot be executed") {
     invalidSqlStatements.foreach { invalidStatement =>
-      val parsedStatement   = SqlStatementReader.readSql(invalidStatement)
-      val extractionResults = TableExtractor.extractTablesFromFlinkRuntime(parsedStatement)
+      val parsedStatement             = SqlStatementReader.readSql(invalidStatement)
+      val sqlStatementExecutionErrors = TablesExtractor.extractTablesFromFlinkRuntime(parsedStatement).invalidValue
 
-      extractionResults.sqlStatementExecutionErrors.size shouldBe 1
-      extractionResults.tableDefinitions shouldBe empty
+      sqlStatementExecutionErrors.size shouldBe 1
     }
   }
 
 }
 
-object TableExtractorTest {
+object TablesExtractorTest {
 
   private val invalidSqlStatements: List[String] = List(
     """|CREATE TABLE testTable