+++ engine/flink/components/table/src/main/scala/pl/touk/nussknacker/engine/flink/table/source/TableSource.scala
@@ -4,8 +4,9 @@ import org.apache.flink.api.common.RuntimeExecutionMode
 import org.apache.flink.streaming.api.datastream.DataStream
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
 import org.apache.flink.table.api.Expressions.$
-import org.apache.flink.table.api.TableEnvironment
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment
+import org.apache.flink.table.api.{DataTypes, Schema, TableEnvironment}
+import org.apache.flink.table.catalog.Column.{ComputedColumn, MetadataColumn, PhysicalColumn}
 import org.apache.flink.types.Row
 import pl.touk.nussknacker.engine.api.component.SqlFilteringExpression
 import pl.touk.nussknacker.engine.api.definition.Parameter
@@ -17,7 +18,6 @@ import pl.touk.nussknacker.engine.api.process.{
   TestWithParametersSupport
 }
 import pl.touk.nussknacker.engine.api.test.{TestData, TestRecord, TestRecordParser}
-import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.flink.api.process.{
   FlinkCustomNodeContext,
   FlinkSourceTestSupport,
@@ -29,6 +29,10 @@ import pl.touk.nussknacker.engine.flink.table.TableComponentProviderConfig.TestD
 import pl.touk.nussknacker.engine.flink.table.TableDefinition
 import pl.touk.nussknacker.engine.flink.table.extractor.SqlStatementReader.SqlStatement
 import pl.touk.nussknacker.engine.flink.table.source.TableSource._
+import pl.touk.nussknacker.engine.flink.table.utils.DataTypesConversions._
+import pl.touk.nussknacker.engine.flink.table.utils.SchemaExtensions._
+
+import scala.jdk.CollectionConverters._
 
 class TableSource(
     tableDefinition: TableDefinition,
@@ -63,15 +67,17 @@ class TableSource(
       }
       .getOrElse(selectQuery)
       // We have to keep elements in the same order as in TypingInfo generated based on TypingResults, see TypingResultAwareTypeInformationDetection
-      .select(tableDefinition.columnNames.sorted.map($): _*)
+      .select(tableDefinition.sourceRowDataType.toLogicalRowTypeUnsafe.getFieldNames.asScala.toList.sorted.map($): _*)
 
     tableEnv.toDataStream(finalQuery)
   }
 
-  override val contextInitializer: ContextInitializer[Row] = new BasicContextInitializer[Row](Typed[Row])
+  override val contextInitializer: ContextInitializer[Row] =
+    new BasicContextInitializer[Row](tableDefinition.sourceRowDataType.getLogicalType.toTypingResult)
 
   override def testParametersDefinition: List[Parameter] =
-    tableDefinition.columns.map(c => Parameter(ParameterName(c.columnName), c.typingResult))
+    fieldsWithoutComputedColumns
+      .map(field => Parameter(ParameterName(field.getName), field.getDataType.getLogicalType.toTypingResult))
 
   override def parametersToTestData(params: Map[ParameterName, AnyRef]): Row = {
     val row = Row.withNames()
@@ -83,23 +89,47 @@ class TableSource(
 
   override def timestampAssignerForTest: Option[TimestampWatermarkHandler[Row]] = None
 
-  override def testRecordParser: TestRecordParser[Row] = (testRecords: List[TestRecord]) =>
-    FlinkMiniClusterTableOperations.parseTestRecords(testRecords, tableDefinition.toFlinkSchema)
-
-  override def generateTestData(size: Int): TestData = testDataGenerationMode match {
-    case TestDataGenerationMode.Random =>
-      FlinkMiniClusterTableOperations.generateRandomTestData(
-        amount = size,
-        schema = tableDefinition.toFlinkSchema
-      )
-    case TestDataGenerationMode.Live =>
-      FlinkMiniClusterTableOperations.generateLiveTestData(
-        limit = size,
-        schema = tableDefinition.toFlinkSchema,
-        sqlStatements = sqlStatements,
-        tableName = tableDefinition.tableName
-      )
+  override def testRecordParser: TestRecordParser[Row] = {
+    val tableDataParserSchema = {
+      val columnsWithMetadataAsPersisted = tableDefinition.schema.getColumns.asScala.map {
+        case p: PhysicalColumn => p
+        case c: ComputedColumn => c
+        case m: MetadataColumn => m.toPhysical
+        case other             => throw new IllegalArgumentException(s"Unknown column type: ${other.getClass}")
+      }.asJava
+      Schema
+        .newBuilder()
+        .fromResolvedSchema(tableDefinition.schema.withColumns(columnsWithMetadataAsPersisted))
+        .build()
+    }
+    (testRecords: List[TestRecord]) =>
+      FlinkMiniClusterTableOperations.parseTestRecords(testRecords, tableDataParserSchema)
   }
+
+  override def generateTestData(size: Int): TestData = {
+    val generateDataSchema = {
+      val dataType = DataTypes.ROW(fieldsWithoutComputedColumns: _*)
+      Schema.newBuilder().fromRowDataType(dataType).build()
+    }
+    testDataGenerationMode match {
+      case TestDataGenerationMode.Random =>
+        FlinkMiniClusterTableOperations.generateRandomTestData(
+          amount = size,
+          schema = generateDataSchema
+        )
+      case TestDataGenerationMode.Live =>
+        FlinkMiniClusterTableOperations.generateLiveTestData(
+          limit = size,
+          schema = generateDataSchema,
+          sqlStatements = sqlStatements,
+          tableName = tableDefinition.tableName
+        )
+    }
+  }
+
+  // We don't want to generate data for computed columns - they will be added during parsing of test data
+  private def fieldsWithoutComputedColumns: List[DataTypes.Field] =
+    tableDefinition.schema.toRowDataTypeFields(c => !c.isInstanceOf[ComputedColumn])
 
 }