+++ docs/Changelog.md
@@ -43,6 +43,8 @@
     * Added ability to run tests on data from file or generated on the spot (the `generated` button)
   * [#6518](https://github.com/TouK/nussknacker/pull/6518) Added user friendly editor for Table Sink output value
     * The editor can be easily switched backed to 'raw' version similarly to the current Kafka Sink
+  * [#6545](https://github.com/TouK/nussknacker/pull/6545) Non-physical columns are properly handled in Table Sink now
+    * User need to provide only persisted column values now - values for computed columns and virtual columns are not required anymore
 
 1.16.2 (18 July 2024)
 -------------------------